[Execution] Inprocess execution works again on windows. [Execution] Renamed *_EXECUTION_ONLY macros to !DISABLE_*_EXECUTION in order to better reflect semantics. [wrapper.setup] Correctly identified all ntdll functions needed. [Execution] Renamed OS-specific sources. All linux sources should have .Linux suffix, and all Windows should have a .Windows suffix. [BinLoader] Wrote ShmMapper as a custom version of InprocMapper.

+++ loader/loader.cpp
@@ -112,7 +112,7 @@ namespace ldr {
 			void *addr = MapMemory(
 				loaderConfig.sections[s].desiredAccess,
 				loaderConfig.sections[s].sectionOffset,
-				loaderConfig.sections[s].mappingSize,
+				(SIZE_T)loaderConfig.sections[s].mappingSize,
 				loaderConfig.sections[s].mappingAddress
 			);
